[Refactor] 프리팹 이름 변경 및 파일 정리
