--- Assets/02.Scripts/FaceCamera.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d1d553938d57448eda26c08b06dc2417
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
