+++ Assets/02.Scripts/Interaction.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3ac8b0b4496be4422b14b7f5869e1250
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
