+++ Assets/02.Scripts/Interaction/FaceCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1d553938d57448eda26c08b06dc2417
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
