--- Assets/02.Scripts/InteractionHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4b14a380cb01142d9a9a17f68e3cb62f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
