--- Assets/07.Prefebs/InteractionUI_(obj_name).prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 10240bdc492f14f30baa98b4fadfac6e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
