+++ Assets/XR/Settings/OpenVRSettings.asset
@@ -18,6 +18,6 @@ MonoBehaviour:
   StereoRenderingMode: 1
   InitializationType: 1
   EditorAppKey: application.generated.unity.myproject.exe
-  ActionManifestFileRelativeFilePath: StreamingAssets\SteamVR\actions.json
+  ActionManifestFileRelativeFilePath: StreamingAssets/SteamVR/actions.json
   MirrorView: 2
   HasCopiedDefaults: 0
